rename files && fix breaking mismatching type in global-charts
