+++ src/app/pages/global/global-initial-state/global-layers.js
