+++ src/app/pages/global/global-initial-state/global-legends.js
