+++ src/app/pages/global/global-initial-state/global-sections.js
